Creating IntegrationCard component for forms modal
