+++ projects/packages/forms/src/blocks/contact-form/components/jetpack-integrations-modal/creative-mail-card.js
@@ -1,0 +1,17 @@
+import { __ } from '@wordpress/i18n';
+import NewsletterIntegrationSettings from '../jetpack-newsletter-integration-settings';
+import IntegrationCard from './integration-card';
+
+const CreativeMailCard = ( { isExpanded, onToggle } ) => {
+	return (
+		<IntegrationCard
+			title={ __( 'Creative Mail', 'jetpack-forms' ) }
+			isExpanded={ isExpanded }
+			onToggle={ onToggle }
+		>
+			<NewsletterIntegrationSettings />
+		</IntegrationCard>
+	);
+};
+
+export default CreativeMailCard;
