+++ projects/packages/forms/src/blocks/contact-form/components/jetpack-integrations-modal/index.js
@@ -1,0 +1,52 @@
+import { Modal } from '@wordpress/components';
+import { useState } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
+import AkismetCard from './akismet-card';
+import CreativeMailCard from './creative-mail-card';
+import JetpackCRMCard from './jetpack-crm-card';
+
+const JetpackIntegrationsModal = ( { isOpen, onClose, attributes, setAttributes } ) => {
+	const [ expandedCards, setExpandedCards ] = useState( {
+		akismet: false,
+		crm: false,
+		creativemail: false,
+	} );
+
+	if ( ! isOpen ) {
+		return null;
+	}
+
+	const toggleCard = cardId => {
+		setExpandedCards( prev => ( {
+			...prev,
+			[ cardId ]: ! prev[ cardId ],
+		} ) );
+	};
+
+	return (
+		<Modal
+			title={ __( 'Manage Integrations', 'jetpack-forms' ) }
+			onRequestClose={ onClose }
+			style={ { width: '700px' } }
+		>
+			<div style={ { padding: '16px', display: 'flex', flexDirection: 'column', gap: '16px' } }>
+				<AkismetCard
+					isExpanded={ expandedCards.akismet }
+					onToggle={ () => toggleCard( 'akismet' ) }
+				/>
+				<JetpackCRMCard
+					isExpanded={ expandedCards.crm }
+					onToggle={ () => toggleCard( 'crm' ) }
+					jetpackCRM={ attributes.jetpackCRM }
+					setAttributes={ setAttributes }
+				/>
+				<CreativeMailCard
+					isExpanded={ expandedCards.creativemail }
+					onToggle={ () => toggleCard( 'creativemail' ) }
+				/>
+			</div>
+		</Modal>
+	);
+};
+
+export default JetpackIntegrationsModal;
