+++ projects/packages/forms/src/blocks/contact-form/edit.js
@@ -226,7 +226,7 @@ function JetpackContactFormEdit( { name, attributes, setAttributes, clientId, cl
 
 					{ isFormModalEnabled && (
 						<PanelBody
-							title={ __( 'Integrations', 'jetpack-forms' ) }
+							title={ __( 'Manage Integrations', 'jetpack-forms' ) }
 							className="jetpack-contact-form__integrations-panel"
 						>
 							<IntegrationPanel attributes={ attributes } setAttributes={ setAttributes } />
